+++ Cassandra.ThriftClient/Abstractions/NetworkTopologyReplicationStrategy.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
 
-using SkbKontur.Cassandra.ThriftClient.Scheme;
+using SkbKontur.Cassandra.ThriftClient.Schema;
 
 namespace SkbKontur.Cassandra.ThriftClient.Abstractions
 {