+++ Cassandra.ThriftClient/Abstractions/SimpleReplicationStrategy.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using System.Globalization;
 
-using SkbKontur.Cassandra.ThriftClient.Scheme;
+using SkbKontur.Cassandra.ThriftClient.Schema;
 
 namespace SkbKontur.Cassandra.ThriftClient.Abstractions
 {