+++ Cassandra.ThriftClient/Schema/ICassandraSchemaActualizer.cs
@@ -1,0 +1,7 @@
+namespace SkbKontur.Cassandra.ThriftClient.Schema
+{
+    public interface ICassandraSchemaActualizer
+    {
+        void ActualizeKeyspaces(KeyspaceSchema[] keyspaceShemas, bool changeExistingKeyspaceMetadata);
+    }
+}