+++ Cassandra.ThriftClient/Schema/KeyspaceConfiguration.cs
@@ -1,6 +1,6 @@
 using SkbKontur.Cassandra.ThriftClient.Abstractions;
 
-namespace SkbKontur.Cassandra.ThriftClient.Scheme
+namespace SkbKontur.Cassandra.ThriftClient.Schema
 {
     public class KeyspaceConfiguration
     {