+++ Cassandra.ThriftClient/Schema/KeyspaceSchema.cs
@@ -1,6 +1,6 @@
-namespace SkbKontur.Cassandra.ThriftClient.Scheme
+namespace SkbKontur.Cassandra.ThriftClient.Schema
 {
-    public class KeyspaceScheme
+    public class KeyspaceSchema
     {
         public string Name { get; set; }
         public KeyspaceConfiguration Configuration { get => configuration ?? new KeyspaceConfiguration(); set => configuration = value; }