+++ Cassandra.ThriftClient/Scheme/ICassandraSchemaActualizer.cs
@@ -1,7 +1,0 @@
-namespace SkbKontur.Cassandra.ThriftClient.Scheme
-{
-    public interface ICassandraSchemaActualizer
-    {
-        void ActualizeKeyspaces(KeyspaceScheme[] keyspaceShemas, bool changeExistingKeyspaceMetadata);
-    }
-}